pyproject.toml
[project]
name = "notes-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.3.0",
    "beanie>=2.0.0",
    "fastapi[all]>=0.120.1",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-jose>=3.5.0",
]
